Fixed the clean and distclean targets to only clean packages built by the all target.
Use a common list of packages to build for use by all of the targets.  A <filelist> element is used since the order in which the packages is built is important.  This element is also valid to use when you want to produce an ordered list of directories.  The subant task works on a list of directories so it isn't necessary to specify file names.